Move unrelated tests
